make it possible to specify compatibility for the project
