+++ stdlib/Pkg/src/Operations.jl
@@ -134,8 +134,9 @@ function collect_project!(ctx::Context, pkg::PackageSpec, path::String, fix_deps
     fix_deps_map[pkg.uuid] = valtype(fix_deps_map)()
     !isfile(project_file) && return false
     project = read_project(project_file)
+    compat = get(project, "compatibility", Dict())
     for (deppkg_name, uuid) in project["deps"]
-        vspec = VersionSpec() # TODO: Update with compatibility from Project
+        vspec = haskey(compat, deppkg_name) ? Types.semver_spec(compat[deppkg_name]) : VersionSpec()
         deppkg = PackageSpec(deppkg_name, UUID(uuid), vspec)
         push!(fix_deps_map[pkg.uuid], deppkg)
     end
@@ -267,20 +268,30 @@ function resolve_versions!(ctx::Context, pkgs::Vector{PackageSpec})::Dict{UUID,V
     for (name::String, uuidstr::String) in ctx.env.project["deps"]
         uuid = UUID(uuidstr)
         uuid_to_name[uuid] = name
-        info = manifest_info(ctx.env, uuid)
-        info == nothing && continue
-        haskey(info, "version") || continue # stdlibs might not have a version
-        ver = VersionNumber(info["version"])
+
         uuid_idx = findfirst(isequal(uuid), uuids)
+        ver = VersionSpec()
         if uuid_idx != nothing
             pkg = pkgs[uuid_idx]
-            if pkg.special_action != PKGSPEC_FREED && get(info, "pinned", false)
-                # This is a pinned package, fix its version
-                pkg.version = ver
+            info = manifest_info(ctx.env, uuid)
+            if info !== nothing && haskey(info, "version") # stdlibs might not have a version
+                ver = VersionNumber(info["version"])
+                    if pkg.special_action != PKGSPEC_FREED && get(info, "pinned", false)
+                        # This is a pinned package, fix its version
+                        pkg.version = ver
+                end
             end
         else
-            push!(pkgs, PackageSpec(name, uuid, ver))
+            pkg = PackageSpec(name, uuid, ver)
+            push!(pkgs, pkg)
         end
+        proj_compat = Types.project_compatibility(ctx, name)
+        v = intersect(pkg.version, proj_compat)
+        if isempty(v)
+            cmderror(string("for package $(pkg.name) intersection between project compatibility $(proj_compat) ",
+                            "and package version $(pkg.version) is empty"))
+        end
+        pkg.version = v
     end
     # construct data structures for resolver and call it
     reqs = Requires(pkg.uuid => VersionSpec(pkg.version) for pkg in pkgs if pkg.uuid ≠ uuid_julia)
